Fix bugs "Camera doesn't follow the player when the player is slowing down"
I'm trying to work on this bug and found this approach, hope it works well.

What I did:

(1) In "CM vcam1" settings, change "Follow" to "Player"

(2) Change "Screen Y" to a value smaller than 0.5, e.g. "0.35", to place the player at the top of the screen

(3) Change "Dead Zone Width" to a value close to 1, e.g. "0.78", to limit the camera following (i.e. the camera will not follow the player until the player reaches the left/right side)

(4) Change "Dead Zone Height" to "0"

There are still some bugs:

(1) when the player reaches the right side, the camera will move a little bit. (This might be related to the bound we set)